Module C

Entry Points

These modules contain the generated types and functions enabling usage of Clipper2 from OCaml.

module Types : sig ... end

module Funcs : sig ... end

Binding Description Functors

Doc comment headers are available to make browsing the bound functions easier in Functions.

module Clipper_c_function_descriptions : sig ... end

module Clipper_c_type_descriptions : sig ... end