Module Make64'.Path

The Clipper2 path type (std::vector of point)

type t = path

Construction / Conversion

val of_list : V.t list -> t

of_list ps

Create a path from the list of 2d points ps.

val to_list : t -> V.t list

to_list t

Convert the path t to a list of 2d points.

val ellipse : ?fn:int -> ?centre:V.t -> V.t -> t

ellipse ?fn ?centre rs

Draw an elliptical path with with the specified xy radii rs centred on the origin, or at the point centre if provided. The number of segments can be set explicitly with fn, otherwise the quality is calculated based on the dimensions.

Access

val length : t -> int

length t

Return the number of points in the path t.

val get_point : t -> int -> V.t

get_point t i

Get the point at index i from the path t.

Boolean Operations

val boolean_op : ?fill_rule:fill_rule -> op:clip_type -> t -> t -> paths

boolean_op ?fill_rule ~op a b

Perform the boolean operation op with the specified fill_rule on the polygons (closed paths) a and b.

val intersect : ?fill_rule:fill_rule -> t -> t -> paths

intersect ?fill_rule a b

Intersect the polygons a and b according to fill_rule. The result includes regions covered by both polygons.

val union : ?fill_rule:fill_rule -> t list -> paths

union ?fill_rule subjects

Union the polygons subjects according to fill_rule. The result includes the regions covered by any of the polygons contained in subjects.

val add : ?fill_rule:fill_rule -> t -> t -> paths

add ?fill_rule a b

union the polygon a and b.

val difference : ?fill_rule:fill_rule -> t -> t -> paths

difference ?fill_rule a b

Difference the polygon b from the polygon a according to fill_rule. The result includes the regions covered by the polygon a, but not the polygon b.

val sub : ?fill_rule:fill_rule -> t -> t -> paths

sub a b

Difference the polygon b from a (alias to difference).

val xor : ?fill_rule:fill_rule -> t -> t -> paths

xor ?fill_rule subjects clips

Perform the exclusive-or boolean operation between the closed paths a and b according to fill_rule. The result includes regions covered by the either a or b, but not both.

val rect_clip : ?closed:bool -> Rect.t -> t -> paths

rect_clip ?closed r t

Intersect the path t with the axis-aligned rectangle r. The path is treated as closed/polygonal by default, but an open path may be clipped by setting ~closed:false.

Offsetting

val inflate : ?join_type:join_type -> ?end_type:end_type -> delta:float -> t -> paths

inflate ?join_type ?end_type ~delta t

Offset the polygon (or open path) t by delta. If t is a closed polygonal path, it's important that end_type is `Polygon (default if not overridden by user's Config). If instead you select one of the open path end types (e.g. `Joined), the polygon's outline will be inflated (example).

with closed paths (polygons), a positive delta specifies how much outer polygon contours will expand and how much inner "hole" contours will contract (and the converse with negative deltas).

with open paths (polylines), including EndType.Join, delta specifies the width of the inflated line.

Caution: offsetting self-intersecting polygons may produce unexpected results.

Minkowski

val minkowski_sum : ?closed:bool -> pattern:t -> t -> t

minkowski_sum ?closed ~pattern t

Apply Minkowski addition of the path pattern to the path t. t is treated as a closed polygon unless otherwise specified.

val minkowski_diff : ?closed:bool -> pattern:t -> t -> t

minkowski_diff ?closed ~pattern t

Apply Minkowski subtraction of the path pattern from the path t. t is treated as a closed polygon unless otherwise specified.

Path Simplification

val simplify : ?closed:bool -> ?eps:float -> t -> t

simplify ?closed ?eps t

Remove extraneous vertices from the path t (similar to ramer_douglas_peucker).

val ramer_douglas_peucker : ?eps:float -> t -> t

ramer_douglas_peucker ?eps t

Applies the Ramer-Douglas-Peucker algorithm to remove extraneous vertices from the path t. Put simply, vertices will be removed if they are less than eps distance from imaginary lines passing through their adjacent vertices.

This function is particularly useful following offsetting (ie inflating/shrinking paths). Offsetting often creates tiny segments that don't improve path quality. Further these tiny segments create angles that are strongly influenced by integer rounding. While these tiny segments are too small to be noticeable following a single offset procedure, they're likely to degrade the quality of subsequent offsets. And they'll also degrade performance. Because of this, it is strongly recommended calling this function after every polygon offset.

val trim_collinear : ?closed:bool -> t -> t

trim_collinear ?closed t

Remove collinear points (that fall on a line drawn between their neighbours) from the path t. The path is treated as closed by default.

val strip_near_equal : ?closed:bool -> ?eps:float -> t -> t

strip_near_equal ?closed ?eps t

Remove adjacent points that are less than eps distance apart from their neighbour from the paths t. The path is treated as closed by default.

val strip_duplicates : ?closed:bool -> t -> t

strip_duplicates ?closed ?eps t

Remove adjacent points that duplicate of their neighbours from the paths t. The path is treated as closed by default.

Transformation

val translate : V.t -> t -> t

translate v t

Translate the path t along the vector v.

Geometry

val area : t -> float

area t

Compute the signed area of the path t.

val bounds : t -> Rect.t

bounds t

Compute the axis-aligned bounding box that contains the path t.

val is_positive : t -> bool

is_positive t

Check if the orientation/winding of the path t is positive.

val point_inside : t -> V.t -> [> `Inside | `OnBorder | `Outside ]

point_inside t p

Determine whether the point p is inside, outside, or on the border of t.