Parameter Make64'.Ctr

Signature of a contour type -- an open or closed path/sequence of points. To be used for construction and destruction of the Clipper2 path type.

type t

contour type

val to_seq : t -> V.t Stdlib.Seq.t

to_seq t converts the contour t to an iterator of points

val of_seq : V.t Stdlib.Seq.t -> t

of_seq vs creates a contour from an iterator points vs