Parameter Make64.V

Signature of a 2d vector type to be used to represent points for the construction and destruction of Clipper2 path types

type t

2d vector type

val v : int64 -> int64 -> t

v x y conststructs a vector from x and y coordinates

val x : t -> int64

x t obtains the x coordinate of the vector t

val y : t -> int64

y t obtains the y coordinate of the vector t