Module MakeD'.Paths

The Clipper2 paths type (std::vector of path)

These lists of contours are not organized hierarchically (by parent-child / outer-hole) relationships, and may include any number of open paths or polygons.

type t = paths

val length : t -> int

length t

Return the number of paths in the list of paths t.

val path_length : t -> int -> int

path_length t i

Return the number of points in the path at index i of t.

val get_path : t -> int -> path

get_path t i

Get the path at index i from t.

val get_point : t -> int -> int -> V.t

get_point t i j

Get the point j from the path at index i in t.

val of_list : V.t list list -> t

of_list ps

Create a path from the list of lists of 2d points ps.

val to_list : t -> V.t list list

to_list t

Convert the paths t to a list of lists of 2d points.

val translate : V.t -> t -> t

translate v t

Translate the paths t along the vector v.

val simplify : ?closed:bool -> ?eps:float -> t -> t

simplify ?closed ?eps t

Remove extraneous vertices from the paths t (similar to ramer_douglas_peucker).

val ramer_douglas_peucker : ?eps:float -> t -> t

ramer_douglas_peucker ?eps t

Applies the Ramer-Douglas-Peucker algorithm to remove extraneous vertices from the paths t. Put simply, vertices will be removed if they are less than eps distance from imaginary lines passing through their adjacent vertices.

This function is particularly useful following offsetting (ie inflating/shrinking paths). Offsetting often creates tiny segments that don't improve path quality. Further these tiny segments create angles that are strongly influenced by integer rounding. While these tiny segments are too small to be noticeable following a single offset procedure, they're likely to degrade the quality of subsequent offsets. And they'll also degrade performance. Because of this, it is strongly recommended calling this function after every polygon offset.

val strip_near_equal : ?closed:bool -> ?eps:float -> t -> t

strip_near_equal ?closed ?eps t

Remove adjacent points that are less than eps distance apart from their neighbour from the paths t. The path is treated as closed by default.

val strip_duplicates : ?closed:bool -> t -> t

strip_duplicates ?closed ?eps t

Remove adjacent points that duplicate of their neighbours from the paths t. The path is treated as closed by default.

val area : t -> float

area t

Compute the summed signed area of all paths in t.

val boolean_op : ?fill_rule:fill_rule -> op:clip_type -> t -> t -> t

boolean_op ?fill_rule ~op subjects clips

Perform the boolean operation op with the specified fill_rule on the paths/polygons subjects with clips (all paths are treated as closed). With the exception of `Union, operations are performed between subjects and clips, not within subjects (e.g. paths in subjects are not differenced from eachother).

val intersect : ?fill_rule:fill_rule -> t -> t -> t

intersect ?fill_rule subjects clips

Intersect the polygons in subjects with the polygons in clips according to fill_rule. The result includes regions covered by both polygons in subjects and clips.

val union : ?fill_rule:fill_rule -> t -> t

union ?fill_rule subjects

Union the polygons subjects according to fill_rule. The result includes the regions covered by any of the polygons contained in subjects.

val add : ?fill_rule:fill_rule -> t -> t -> t

add ?fill_rule a b

union all of the polygons in a and b.

val difference : ?fill_rule:fill_rule -> t -> t -> t

difference ?fill_rule subjects clips

Difference the polygons in clips from subjects according to fill_rule. The result includes the regions covered by the subjects, but not the clips polygons.

val sub : ?fill_rule:fill_rule -> t -> t -> t

sub a b

Difference the polygons in b from a (alias to difference).

val xor : ?fill_rule:fill_rule -> t -> t -> t

xor ?fill_rule subjects clips

Perform the exclusive-or boolean operation between the closed paths subjects and clips according to fill_rule. The result includes regions covered by the subjects or clips polygons, but not both.

val inflate : ?join_type:join_type -> ?end_type:end_type -> delta:float -> t -> t

inflate ?join_type ?end_type ~delta t

Offset the polygons (and/or open paths) in t by delta. If you intend to inflate polygons (closed paths), it's important that end_type is `Polygon (default if not overridden by user's Config). If instead you select one of the open path end types (e.g. `Joined), the polygon's outline will be inflated (example).

with closed paths (polygons), a positive delta specifies how much outer polygon contours will expand and how much inner "hole" contours will contract (and the converse with negative deltas).

with open paths (polylines), including EndType.Join, delta specifies the width of the inflated line.

Caution: offsetting self-intersecting polygons may produce unexpected results.

val of_poly : P.t -> t

of_poly p

Create a paths from a polygon p.

val of_polys : P.t list -> t

of_polys ps

Create a paths from a list of polygons ps.

val to_polys : ?fill_rule:fill_rule -> t -> P.t list

to_polys ?fill_rule t

Extract a list of non-overlapping polygons from the set of paths t. This involves a clipper union operation tracking the parent-child (outline-hole) relationships of the paths, thus fill_rule can be provided to override the default rule if desired.