Parameter MakeD'.P

Signature of a polygon type -- an outer path and zero or more inner paths (holes). To be used for construction and destruction of Clipper2 paths types.

type t

polygon type

val to_seq : t -> V.t Stdlib.Seq.t Stdlib.Seq.t

to_seq t converts the polygon t to an iterator of iterators of points

val of_seq : V.t Stdlib.Seq.t Stdlib.Seq.t -> t

of_seq vs creates a polygon from the iterator of iterators of points vs