Parameter MakeD'._

Configuration parameters allowing user control Clipper2 interface defaults

val fill_rule : [ `EvenOdd | `NonZero | `Positive | `Negative ] option

The default filling rule used by the clipping algorithm (for boolean operations)

val join_type : [ `Square | `Round | `Miter ] option

The default treatment of corners when offsetting paths

val end_type : [ `Polygon | `Joined | `Butt | `Square | `Round ] option

The default for whether paths are treated as closed (polygons), or as open paths (and if so, what open path strategy to take)

val precision : int option

The number of significant decimal places to use during numerical operations (if using the float interface). Can be up to 8.

val eps : float option

The default epsilon value to use for relevant functions.