Module type Clipper.Poly

Signature of a polygon type -- an outer path and zero or more inner paths (holes). To be used for construction and destruction of Clipper2 paths types.

type v

2d vector type representing points -- will be destructed (type declaration not required in user supplied module)

type t

polygon type

val to_list : t -> v list list

to_list t converts the polygon t to a list of lists of points

val of_list : v list list -> t

of_list vs creates a polygon from the list of lists of points vs