Module S.Path

The Clipper2 path type (std::vector of point)

type t = path

val length : t -> int

length t

Return the number of points in the path t.

val get_point : t -> int -> v

get_point t i

Get the point at index i from the path t.

val of_list : v list -> t

of_list ps

Create a path from the list of 2d points ps.

val to_list : t -> v list

to_list t

Convert the path t to a list of 2d points.

val ellipse : ?fn:int -> ?centre:v -> v -> t

ellipse ?fn ?centre rs

Draw an elliptical path with with the specified xy radii rs centred on the origin, or at the point centre if provided. The number of segments can be set explicitly with fn, otherwise the quality is calculated based on the dimensions.

val translate : v -> t -> t

translate v t

Translate the path t along the vector v.

val simplify : ?closed:bool -> ?eps:float -> t -> t

simplify ?closed ?eps t

Remove extraneous vertices from the path t (similar to ramer_douglas_peucker).

val ramer_douglas_peucker : ?eps:float -> t -> t

ramer_douglas_peucker ?eps t

Applies the Ramer-Douglas-Peucker algorithm to remove extraneous vertices from the path t. Put simply, vertices will be removed if they are less than eps distance from imaginary lines passing through their adjacent vertices.

This function is particularly useful following offsetting (ie inflating/shrinking paths). Offsetting often creates tiny segments that don't improve path quality. Further these tiny segments create angles that are strongly influenced by integer rounding. While these tiny segments are too small to be noticeable following a single offset procedure, they're likely to degrade the quality of subsequent offsets. And they'll also degrade performance. Because of this, it is strongly recommended calling this function after every polygon offset.

val trim_collinear : ?closed:bool -> t -> t

trim_collinear ?closed t

Remove collinear points (that fall on a line drawn between their neighbours) from the path t. The path is treated as closed by default.

val strip_near_equal : ?closed:bool -> ?eps:float -> t -> t

strip_near_equal ?closed ?eps t

Remove adjacent points that are less than eps distance apart from their neighbour from the paths t. The path is treated as closed by default.

val strip_duplicates : ?closed:bool -> t -> t

strip_duplicates ?closed ?eps t

Remove adjacent points that duplicate of their neighbours from the paths t. The path is treated as closed by default.

val point_inside : t -> v -> [> `Inside | `OnBorder | `Outside ]

point_inside t p

Determine whether the point p is inside, outside, or on the border of t.

val area : t -> float

area t

Compute the signed area of the path t.

val is_positive : t -> bool

is_positive t

Check if the orientation/winding of the path t is positive.