Module Clpr.Offset

This class provides a mixed interface (float and int64). Unlike the purpose built ClipperD.t it does not perform handle the conversion of incoming doubles to int64 using a precision set by the user -- they are simply truncated. Thus if you are working with floats, it is recommended that you make use of the helper function PathsD.inflate at this time.

type t

val make : ?miter_limit:float -> ?arc_tolerance:float -> ?preserve_collinear:bool -> ?reverse_solution:bool -> unit -> t

val add_pathd : ?join_type:join_type -> ?end_type:end_type -> t -> PathD.t -> unit

val add_pathsd : ?join_type:join_type -> ?end_type:end_type -> t -> PathsD.t -> unit

val add_path64 : ?join_type:join_type -> ?end_type:end_type -> t -> Path64.t -> unit

val add_paths64 : ?join_type:join_type -> ?end_type:end_type -> t -> Paths64.t -> unit

val execute : delta:float -> t -> Paths64.t

val get_miter_limit : t -> float

val set_miter_limit : t -> float -> unit

val get_arc_tolerance : t -> float

val set_arc_tolerance : t -> float -> unit

val get_preserve_collinear : t -> bool

val set_preserve_collinear : t -> bool -> unit

val get_reverse_solution : t -> bool

val set_reverse_solution : t -> bool -> unit

val error_code : t -> int

val clear : t -> unit