Module Clpr.PolyTree64

PolyTree64.t is a read-only data structure that receives solutions from clipping operations. It's an alternative to the Paths64.t data structure which also receives solutions. However the principle advantage of PolyTree64.t over Paths64.t is that it also represents the parent-child relationships of the polygons in the solution (where a parent's polygon will contain all its children polygons).

The PolyTree64.t object that's to receive a clipping solution is passed as a parameter to Clipper64.execute. When the clipping operation finishes, this object will be populated with data representing the clipped solution.

A PolyTree64.t object is a container for any number of PolyTree64.t child objects, each representing a single polygon contour. The top-level descendents of a PolyTree64.t solution will always be outer polygon contours. Children may in turn contain their own children to any level of nesting. Children of outer polygon contours will always represent holes, and children of holes will always represent nested outer polygon contours.

PolyTree64.t will never contain open paths since open paths can't contain paths. When clipping open paths, these will always be represented in solutions via a separate Paths64.t structure.

type t

val make : ?parent:t -> unit -> t

val clear : t -> unit

val parent : t -> t option

val child : t -> int -> t

val level : t -> int

val count : t -> int

val is_hole : t -> bool

val area : t -> float

val polygon : t -> Path64.t

val decompose : (Path64.t -> 'a) -> t -> 'a Stdlib.Seq.t Stdlib.Seq.t