Module type Clipper.S

The signature of Clipper2 binding modules produced by the provided Make functors

type v

2d vector type representing points

type poly

polygon type -- outer path and zero or more inner paths (holes)

type path

The Clipper2 Path type (std::vector of point)

type paths

The Clipper2 Paths type (std::vector of path)

type clip_type = [

| `None

| `Intersection
(*
AND -- regions covered by both subject and clip polygons
*)

| `Union
(*
OR -- regions covered by subject or clip polygons, or both polygons
*)

| `Difference
(*
NOT -- regions covered by subject, but not clip polygons
*)

| `Xor
(*
exclusive or -- regions covered by subject or clip polygons, but not both
*)

]

Clipping types for boolean operations. See Clipper2's docs for visual demonstrations.

type fill_rule = [

| `EvenOdd
(*
only odd numbered sub-regions are filled
*)

| `NonZero
(*
non-zero sub-regions are filled
*)

| `Positive
(*
only sub-regions with winding counts > 0 are filled
*)

| `Negative
(*
only sub-regions with winding counts < 0 are filled
*)

]

Filling rules used by the clipping algorithm for boolean operations. See Clipper2's docs for a detailed explanation of how they differ).

type join_type = [

| `Square
(*
squaring applied uniformally at all joins where the internal join angle is less than 90 degrees. The squared edg will be at exactly the offset distance from the join vertex
*)

| `Round
(*
rounding is appliedto all joins that have convex external angles, and it maintains the exact offset distance from the join vertex
*)

| `Miter of float option
(*
there's a necessary limit to mitered joins (to avoid narrow angled joins producing excessively long and narrow spikes)). The optional limit sets the maximum distance in multiples of the delta specified for the offsetting operation (default is 2, which is the minimum allowed).
*)

]

Defines the treatment of corners when offsetting paths. Visual examples are available in the Clipper2 docs.

type end_type = [

| `Polygon
(*
paths are assumed to be closed and treated as polygons
*)

| `Joined
(*
ends are joined and the paths are treated as polylines
*)

| `Butt
(*
ends are squared off without any extrusion
*)

| `Square
(*
ends extend the offset amount while being squared off
*)

| `Round
(*
ends extend the offset amount while being rounded off
*)

]

Sets whether paths are treated as closed (`Polygon) when offsetting or open (and how to do so, if so). Visual examples are available in the Clipper2 docs.

module Path : sig ... end

The Clipper2 path type (std::vector of point)

module Paths : sig ... end

The Clipper2 paths type (std::vector of path)